setup/requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.12
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argostranslate==1.9.6
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.4
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
compel==2.0.2
contourpy==1.2.1
cryptography==3.4.8
ctranslate2==4.3.1
cycler==0.12.1
Cython==3.0.10
dbus-python==1.2.18
deep-translator==1.11.4
diffusers==0.29.1
distro==1.7.0
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.1
fairscale==0.4.13
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.43
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
importlib-metadata==4.6.4
insightface==0.7.3
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lark==1.1.9
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lightning-utilities==0.11.2
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matrix-client==0.4.0
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.16.1
onnxruntime-gpu==1.18.0
openai==1.35.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
piexif==1.1.3
pillow==10.2.0
prettytable==3.10.0
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==3.1.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
sacremoses==0.0.53
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.1
segment-anything==1.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
spandrel==0.3.4
stanza==1.1.1
sympy==1.12
threadpoolctl==3.5.0
tifffile==2024.6.18
timm==1.0.7
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchmetrics==1.4.0.post0
torchsde==0.2.6
torchvision==0.18.0+cu121
tqdm==4.66.4
trampoline==0.1.2
transformers==4.41.2
triton==2.3.0
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.42
ultralytics-thop==2.0.0
urllib3==1.26.19
wadllib==1.3.6
wcwidth==0.2.13
xformers==0.0.26.post1
yarl==1.9.4
zipp==1.0.0
